Scrollbar smooth y resuelto conflictos del niki

+++ MarioBros/MarioClone/src/character/CharacterEntity.java
@@ -5,6 +5,6 @@ import game.LogicalEntity;
 public interface CharacterEntity extends LogicalEntity {
     public int getLives();
     public boolean isInvincible();
-    public int getSpeed();
+    public float getSpeed();
 }
